Initial commit to github
This is the initial commit to github. The working project (for now) is
in the SFML folder. The project rely to some functionalities of the SFML
libraries - sfml ver. 2.4.0. You must download and manually set up this
version of SFML in the project properties in order the code to be built
properly.
Made so far - main.cpp added where a test window is displayed which when
it become inactive add an tray icon.