Power scheme change functionality added.
Now the program is able to track for changes on the power scheme. When
the system run from AC power to battery then a popup window appears
showing a warning.
Also when left click on the icon in the system tray then a popup shows
with info on the current power scheme and the precentage of the battery.
Added an early version of an imgae for replacing the one of the icon in
the system tray.